Scheduler: move nodedb.Node to a new internaltypes package
